refactor(pinned events): move `pin_event`/`unpin_event` from the Timeline to the Room
These make sense in general, and they will help getting rid of one of
the `PinnedEventsRoom` trait methods in a subsequent PR.
