+++ crates/matrix-sdk-ui/src/timeline/mod.rs
@@ -39,7 +39,6 @@ use matrix_sdk::{
     send_queue::{RoomSendQueueError, SendHandle},
 };
 use mime::Mime;
-use pinned_events_loader::PinnedEventsRoom;
 use ruma::{
     EventId, OwnedEventId, OwnedTransactionId, UserId,
     api::client::receipt::create_receipt::v3::ReceiptType,
@@ -48,12 +47,9 @@ use ruma::{
         poll::unstable_start::{NewUnstablePollStartEventContent, UnstablePollStartEventContent},
         receipt::{Receipt, ReceiptThread},
         relation::Thread,
-        room::{
-            message::{
-                Relation, RelationWithoutReplacement, ReplyWithinThread,
-                RoomMessageEventContentWithoutRelation, TextMessageEventContent,
-            },
-            pinned_events::RoomPinnedEventsEventContent,
+        room::message::{
+            Relation, RelationWithoutReplacement, ReplyWithinThread,
+            RoomMessageEventContentWithoutRelation, TextMessageEventContent,
         },
     },
     room_version_rules::RoomVersionRules,
@@ -828,58 +824,6 @@ impl Timeline {
                 self.room().set_unread_flag(false).await?;
             }
 
-            Ok(false)
-        }
-    }
-
-    /// Adds a new pinned event by sending an updated `m.room.pinned_events`
-    /// event containing the new event id.
-    ///
-    /// This method will first try to get the pinned events from the current
-    /// room's state and if it fails to do so it'll try to load them from the
-    /// homeserver.
-    ///
-    /// Returns `true` if we pinned the event, `false` if the event was already
-    /// pinned.
-    pub async fn pin_event(&self, event_id: &EventId) -> Result<bool> {
-        let mut pinned_event_ids = if let Some(event_ids) = self.room().pinned_event_ids() {
-            event_ids
-        } else {
-            self.room().load_pinned_events().await?.unwrap_or_default()
-        };
-        let event_id = event_id.to_owned();
-        if pinned_event_ids.contains(&event_id) {
-            Ok(false)
-        } else {
-            pinned_event_ids.push(event_id);
-            let content = RoomPinnedEventsEventContent::new(pinned_event_ids);
-            self.room().send_state_event(content).await?;
-            Ok(true)
-        }
-    }
-
-    /// Removes a pinned event by sending an updated `m.room.pinned_events`
-    /// event without the event id we want to remove.
-    ///
-    /// This method will first try to get the pinned events from the current
-    /// room's state and if it fails to do so it'll try to load them from the
-    /// homeserver.
-    ///
-    /// Returns `true` if we unpinned the event, `false` if the event wasn't
-    /// pinned before.
-    pub async fn unpin_event(&self, event_id: &EventId) -> Result<bool> {
-        let mut pinned_event_ids = if let Some(event_ids) = self.room().pinned_event_ids() {
-            event_ids
-        } else {
-            self.room().load_pinned_events().await?.unwrap_or_default()
-        };
-        let event_id = event_id.to_owned();
-        if let Some(idx) = pinned_event_ids.iter().position(|e| *e == *event_id) {
-            pinned_event_ids.remove(idx);
-            let content = RoomPinnedEventsEventContent::new(pinned_event_ids);
-            self.room().send_state_event(content).await?;
-            Ok(true)
-        } else {
             Ok(false)
         }
     }
