+++ crates/matrix-sdk/tests/integration/room/mod.rs
@@ -6,6 +6,7 @@ mod common;
 mod joined;
 mod left;
 mod notification_mode;
+mod pinned_events;
 mod spaces;
 mod tags;
 mod thread;
